requirements.txt
asgiref==3.7.2
Django==4.2.2
joblib==1.2.0
numpy==1.25.0
pyserial==3.5
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
threadpoolctl==3.1.0
